refactor visual editor scripts to prepare for growth
