+++ public/css/grapes-component.css
@@ -1,0 +1,82 @@
+
+.CodeMirror * {
+    text-transform: none;
+    font-family: Arial, monospace;
+    font-size: 16px;
+    /*min-height: 450px;
+    margin-bottom: 8px;*/
+  }
+  .icons-flex {
+    background-size: 70% 65% !important;
+    height: 15px;
+    width: 17px;
+    opacity: 0.9;
+  }
+  .icon-dir-row {
+    background: url("./img/flex-dir-row.png") no-repeat center;
+  }
+  .icon-dir-row-rev {
+    background: url("./img/flex-dir-row-rev.png") no-repeat center;
+  }
+  .icon-dir-col {
+    background: url("./img/flex-dir-col.png") no-repeat center;
+  }
+  .icon-dir-col-rev {
+    background: url("./img/flex-dir-col-rev.png") no-repeat center;
+  }
+  .icon-just-start{
+   background: url("./img/flex-just-start.png") no-repeat center;
+  }
+  .icon-just-end{
+   background: url("./img/flex-just-end.png") no-repeat center;
+  }
+  .icon-just-sp-bet{
+   background: url("./img/flex-just-sp-bet.png") no-repeat center;
+  }
+  .icon-just-sp-ar{
+   background: url("./img/flex-just-sp-ar.png") no-repeat center;
+  }
+  .icon-just-sp-cent{
+   background: url("./img/flex-just-sp-cent.png") no-repeat center;
+  }
+  .icon-al-start{
+   background: url("./img/flex-al-start.png") no-repeat center;
+  }
+  .icon-al-end{
+   background: url("./img/flex-al-end.png") no-repeat center;
+  }
+  .icon-al-str{
+   background: url("./img/flex-al-str.png") no-repeat center;
+  }
+  .icon-al-center{
+   background: url("./img/flex-al-center.png") no-repeat center;
+  }
+
+   [data-tooltip]::after {
+     background: rgba(51, 51, 51, 0.9);
+   }
+
+   .gjs-pn-commands {
+     min-height: 40px;
+   }
+
+   #gjs-sm-float {
+      display: none;
+   }
+
+   .gjs-logo-version {
+     background-color: #756467;
+   }
+
+  .gjs-pn-btn.gjs-pn-active {
+    box-shadow: none;
+  }
+
+  .grp-handler-close {
+    background-color: transparent;
+    color: #ddd;
+  }
+
+  .grp-handler-cp-wrap {
+    border-color: transparent;
+  }
