+++ resources/views/livewire/site-page-editor.blade.php
@@ -9,7 +9,7 @@
 
             <button wire:click="create()" class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded my-3">Create New Site Page</button>
             @if($isOpen)
-                @include('sitepage.create-or-edit')
+                @include('sitepage.create-or-edit', ['team_selection' => $team_selection])
             @endif          
             @if($isVisualEditorOpen)
                 @include('sitepage.visual-editor')
