+++ resources/views/sites/my-site-editor.blade.php
@@ -1,6 +1,6 @@
 <x-app-layout> 
     <div>
-        @livewire('site.create-or-edit',['show_modal' => 0, 'site' => $site,'user'=>$user,'team'=>$team ])
+        @livewire('site.create-or-edit',['show_modal' => 0, 'site' => $site,'user'=>$user,'team'=>$team, 'team_selection'=>$team_selection ])
     </div>
     <div>
     @livewire('site-page-editor',['siteid'=>$site->id]);
