+++ resources/views/sites/show.blade.php
@@ -15,7 +15,7 @@
 
         <div class="max-w-7xl mx-auto py-10 sm:px-6 lg:px-8">
             
-            @livewire('site-editor');
+            @livewire('site-editor',['team_selection' => $team_selection]);
         </div>
     </div>
